Remove unused resources and code
