--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/NativeCameraDemo.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 431a2670c32323440b18bb63ae581fdc
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
