--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/RestTest.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f8a6cb6a9f9275647b04a7f7e1cb17a6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
