--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/UnimgpickerDemo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4ada03948082e0748adb02295306a7e0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
