--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/UnimgpickerDemo/Picker.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c4fa35b147cc5cb4d9ae35310a6c85a3
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
