--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/UnimgpickerDemo/PickerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 757b355315336e345bb7d960d88c6667
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
