--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/UnimgpickerDemo/Rotator.cs
@@ -1,16 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-namespace Kakera
-{
-    public class Rotator : MonoBehaviour
-    {
-        [SerializeField]
-        private Vector3 rotationVector;
-
-        void Update()
-        {
-            transform.Rotate(rotationVector * Time.deltaTime);
-        }
-    }
-}
