--- Prototype/Unity/Superdoku/Assets/Scenes/Debug/UnimgpickerDemo/Rotator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4b7beac40494ab748934761892a5eab9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
