--- Prototype/Unity/Superdoku/Assets/Scripts/Demo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 42674999aaa1e724399c5c29ce62ef29
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
