--- Prototype/Unity/Superdoku/Assets/Scripts/Demo/NativeCameraDemo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 820aef8b8fd32c94f81b3fbe013112b3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
