--- Prototype/Unity/Superdoku/Assets/Scripts/Utils/ObjectDragger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c2bf88df7575d348bb3cabccb06678e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
