--- Prototype/Unity/Superdoku/Assets/Scripts/Utils/ObjectResizer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea50837e8f8ec534193fd213943a96dd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
