--- Prototype/Unity/Superdoku/Assets/Scripts/Utils/ResizeHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e07554d3fc9fd14292fdf7812cde495
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
