--- Prototype/Unity/Superdoku/Assets/Scripts/WebCamPhotoCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b0fc566f3a9625419975c1c5ec17a46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
